Add 0 boilerplate sync and async client objects
With the current approach, for every API call, the user has to import a
a module from a tag package and call the `asyncio` or `sync` method with
the "client=client" argument. This patch adds a wrapper around tag
packages and two wrapper `Client` objects to spare the need for that
boilerplate code.

Check this issue for more information:
https://github.com/triaxtec/openapi-python-client/issues/224

`base_url` can be omitted during client initialization
if `SPEC_TITLE_BASE_URL` is set. Object-oriented clients
have also been renamed to include the spec title.